[SonicUnleashedHUD] Use int for ButtonType

+++ DLLModsSource/SonicGenerations/SonicUnleashedHUD/UnleashedHUD/Configuration.h
@@ -35,7 +35,7 @@ public:
 	/// </summary>
 	static WindowTitleRegion windowTitleRegion;
 
-	static enum class ButtonType : byte
+	static enum class ButtonType : int
 	{
 		X360,
 		XSX,
